+++ Dev/Dev2.Data/PathOperations/PerformIntegerIOOperation.cs
@@ -8,10 +8,10 @@
 *  @license GNU Affero General Public License <http://www.gnu.org/licenses/agpl-3.0.html>
 */
 
-using Dev2.Common.Interfaces.Wrappers;
-using Dev2.Data.Interfaces;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
+using Dev2.Common.Interfaces.Wrappers;
+using Dev2.Data.Interfaces;
 
 namespace Dev2.Data.PathOperations
 {